mybokeh/requirements.txt
bokeh==0.12.0
Django==1.9.7
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.1
python-dateutil==2.5.3
PyYAML==3.11
requests==2.10.0
six==1.10.0
tornado==4.3